+++ Controller/productController.go
@@ -1,8 +1,7 @@
-package API
+package Controller
 
 import (
 	"encoding/json"
-	"fmt"
 	"github.com/gorilla/mux"
 	"github.com/kartikeya/product_catalog_DIY/Database"
 	"github.com/kartikeya/product_catalog_DIY/Model"
@@ -49,7 +48,6 @@ func GetProductById(w http.ResponseWriter, r *http.Request) {
 	w.Header().Set("Content-type", "application/json")
 	params := mux.Vars(r)
 	var product Model.Product
-	fmt.Println("printing product----")
 	Database.DB.First(&product, params["id"])
 	//fmt.Println(product)
 	json.NewEncoder(w).Encode(product)
