+++ go.mod
@@ -3,6 +3,7 @@ module github.com/kartikeya/product_catalog_DIY
 go 1.18
 
 require (
+	github.com/golang/mock v1.6.0
 	github.com/gorilla/mux v1.8.0
 	gorm.io/driver/postgres v1.3.4
 	gorm.io/gorm v1.23.4
